requirements.txt
pytest
pytest-bdd
pytest-selenium